# topdown: Remove unnecessary allocations for binding undos