--- a/topdown/bindings.go
+++ b/topdown/bindings.go
@@ -12,9 +12,8 @@ import (
 )
 
 type undo struct {
-	k    *ast.Term
-	u    *bindings
-	next *undo
+	k *ast.Term
+	u *bindings
 }
 
 func (u *undo) Undo() {
@@ -27,7 +26,6 @@ func (u *undo) Undo() {
 		return
 	}
 	u.u.delete(u.k)
-	u.next.Undo()
 }
 
 type bindings struct {
@@ -123,12 +121,13 @@ func (u *bindings) plugNamespaced(a *ast.Term, caller *bindings) *ast.Term {
 	return a
 }
 
-func (u *bindings) bind(a *ast.Term, b *ast.Term, other *bindings) *undo {
+func (u *bindings) bind(a *ast.Term, b *ast.Term, other *bindings, und *undo) {
 	u.values.Put(a, value{
 		u: other,
 		v: b,
 	})
-	return &undo{a, u, nil}
+	und.k = a
+	und.u = u
 }
 
 func (u *bindings) apply(a *ast.Term) (*ast.Term, *bindings) {